+++ src/methods/sendChatbotRequest/types.ts
@@ -1,0 +1,7 @@
+export type SendChatbotRequestParams = {
+  chatbotContext: { role: string; content: string }[];
+  enhancedUserMessage: string;
+  systemInstructions: string;
+  userContext?: any[];
+  openAIKey: string;
+};
