+++ tsconfig.json
@@ -7,8 +7,7 @@
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
     "allowSyntheticDefaultImports": true,
-    "allowImportingTsExtensions": true,
-    "noEmit": true,
+    "noEmit": false,
     "skipLibCheck": true,
     "types": ["node", "jest"],
     "outDir": "./dist",
@@ -18,17 +17,17 @@
     "noFallthroughCasesInSwitch": true,
     "strictPropertyInitialization": false,
     "declaration": true,
+    "declarationDir": "./dist/types",
     "baseUrl": ".",
     "paths": {
-      "@src/*": ["src/*"],
-      "@utils/*": ["src/utils/*"],
+      "@methods/*": ["src/methods/*"],
       "@config": ["src/config.ts"],
-      "@general/*": ["src/general/*"]
+      "@src/*": ["src/*"],
     }
-  },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist"],
-  "ts-node": {
+	},
+	"include": ["src/**/*"],
+	"exclude": ["node_modules", "dist"],
+	"ts-node": {
     "require": ["tsconfig-paths/register"]
   }
 }
